fix(DSWx-S1): apply dedupe logic to ES query in download job
